+++ main.py
@@ -17,7 +17,7 @@ from telegram.ext import (
     filters,
 )
 
-from hiroshi.config import telegram_settings, application_settings
+from hiroshi.config import application_settings, telegram_settings
 from hiroshi.services.bot import (
     handle_available_providers_options,
     handle_prompt,
@@ -165,11 +165,11 @@ class HiroshiBot:
         # )
         app.add_error_handler(self.error_handler)
         if not app.job_queue:
-            logger.error('Application job queue was shut down or never started.')
+            logger.error("Application job queue was shut down or never started.")
         else:
-            app.job_queue.run_repeating(callback=run_monitoring,
-                                        interval=application_settings.monitoring_frequency_call,
-                                        first=0.0)
+            app.job_queue.run_repeating(
+                callback=run_monitoring, interval=application_settings.monitoring_frequency_call, first=0.0
+            )
 
         app.run_polling()